+++ RedstoneScript/Models/Signals/BreakSignal.cs
@@ -1,0 +1,6 @@
+namespace RedstoneScript.Interpreter.Signals;
+
+public class BreakSignal : Exception
+{
+
+}